refactor component scss files
