+++ src/main.scss
@@ -150,5 +150,3 @@ $contentM: 160px;
     height: 100%;
   }
 }
-
-@import './support/debug-log';
